Update delete button and add dialogs

+++ semana-7/crud-pokemon/js/index.js
@@ -49,9 +49,41 @@ const readPokemons = () => {
 };
 
 const deletePokemon = (index) => {
-  pokemons.splice(index, 1);
-  localStorage.setItem(POKEMONS_CRUD_DATA, JSON.stringify(pokemons));
-  readPokemons();
+  const swalWithBootstrapButtons = Swal.mixin({
+    customClass: {
+      confirmButton: 'btn btn-success mx-1',
+      cancelButton: 'btn btn-danger mx-1'
+    },
+    buttonsStyling: false
+  });
+  swalWithBootstrapButtons.fire({
+    title: '¿Está seguro?',
+    text: "¡No podrás revertir esto!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonText: '¡Sí, elimínalo!',
+    cancelButtonText: '¡No, cancélalo!',
+    reverseButtons: true
+  }).then((result) => {
+    if (result.isConfirmed) {
+      pokemons.splice(index, 1);
+      localStorage.setItem(POKEMONS_CRUD_DATA, JSON.stringify(pokemons));
+      readPokemons();
+      swalWithBootstrapButtons.fire(
+        '¡Eliminado!',
+        'Tu registro ha sido eliminado.',
+        'success'
+      );
+    } else if (
+      result.dismiss === Swal.DismissReason.cancel
+    ) {
+      swalWithBootstrapButtons.fire(
+        'Cancelado',
+        'Tu registro está seguro',
+        'error'
+      );
+    }
+  });
 };
 
 const documentReady = () => {
